feat: remove crossed dependency with server.ts; move doc format req
