+++ server/src/constants.ts
@@ -1,0 +1,4 @@
+import { createConnection, IConnection, IPCMessageReader, IPCMessageWriter } from 'vscode-languageserver';
+
+// Create a connection for the server. The connection uses Node's IPC as a transport
+export const connection: IConnection = createConnection(new IPCMessageReader(process), new IPCMessageWriter(process));
