+++ server/src/features/index.ts
@@ -2,3 +2,4 @@ export * from './CompletionResolveRequestHandler';
 export * from './DocumentSymbolRequestHandler';
 export * from './FoldingRangesRequestHandler';
 export * from './CodeLensRequestHandler';
+export * from './DocumentFormattingRequestHandler';
