+++ server/src/visitors/YmlParsingErrorListener.ts
@@ -1,10 +1,8 @@
 import { ANTLRErrorListener, RecognitionException, Recognizer, Token } from 'antlr4ts';
-import { Diagnostic, Range } from 'vscode-languageserver';
-
-import { parsingIssueSeverityLevel } from '../server';
+import { Diagnostic, DiagnosticSeverity, Range } from 'vscode-languageserver';
 
 export class YmlParsingErrorListener implements ANTLRErrorListener<Token> {
-    constructor(public diagnostics: Diagnostic[]) {}
+    constructor(public diagnostics: Diagnostic[], public severityLevel: DiagnosticSeverity) {}
 
     public syntaxError(
         // tslint:disable-next-line: variable-name
@@ -24,7 +22,7 @@ export class YmlParsingErrorListener implements ANTLRErrorListener<Token> {
         const diagnostic = Diagnostic.create(
             Range.create(currentEditorLine, charPositionInLine, currentEditorLine, endPosition),
             msg,
-            parsingIssueSeverityLevel,
+            this.severityLevel,
         );
         this.diagnostics.push(diagnostic);
     }
